Move `Variable<T>` from 'lib.rs' into separate file 'variable.rs'
